restructure directories
Put components used in a view nearer to the view.
