+++ src/components/routing/guards/GameGuard.tsx
@@ -7,9 +7,9 @@ import { toIri } from "../../../util/toIri"
 import { ApiError, EntityModelGame, Game } from "../../../generated"
 import { Loading } from "../../ui/Loading"
 import { delay } from "../../../util/timeout"
-import { Room } from "../../views/Room"
-import GameView from "../../views/GameView"
-import { GameSummary } from "../../views/GameSummary"
+import { Room } from "../../views/game/room/Room"
+import GameView from "../../views/game/game/GameView"
+import { GameSummary } from "../../views/game/gamesummary/GameSummary"
 import BaseContainer from "../../ui/BaseContainer"
 import SendIcon from "@mui/icons-material/Send"
 import Button from "@mui/material/Button"
