+++ src/components/routing/routers/Paths.ts
@@ -3,5 +3,4 @@ export enum Paths {
   CREATE_PLAYER = "/createPlayer",
   LOBBY = "/lobby",
   GAME = "/game",
-  SANDBOX = "/sandbox",
 }
