--- src/components/views/Announcement.tsx
@@ -1,12 +1,0 @@
-import React from "react"
-import BaseContainer from "../ui/BaseContainer"
-
-const Announcement = (props) => {
-  return (
-    <BaseContainer>
-      <h1>Announcement page</h1>
-    </BaseContainer>
-  )
-}
-
-export default Announcement
