--- src/components/views/Sandbox.tsx
@@ -1,11 +1,0 @@
-import React from "react"
-
-const Sandbox = (props) => {
-  return (
-    <div>
-      <p>Sandbox</p>
-    </div>
-  )
-}
-
-export default Sandbox
