--- src/components/views/Scoreboard.tsx
@@ -1,17 +1,0 @@
-import React from "react"
-import BaseContainer from "../ui/BaseContainer"
-
-/**
- * After a complete game is played this view gone display
- * @param props
- * @returns
- */
-const Scoreboard = (props) => {
-  return (
-    <BaseContainer>
-      <h1>Scoreboard page</h1>
-    </BaseContainer>
-  )
-}
-
-export default Scoreboard
