+++ src/components/views/game/game/GameView.tsx
@@ -1,17 +1,17 @@
 import React from "react"
 import { useNavigate } from "react-router-dom"
-import { Paths } from "../routing/routers/Paths"
-import { usePlayers } from "../../hooks/api/usePlayers"
-import { useCurrentGame } from "../../hooks/api/useCurrentGame"
+import { Paths } from "../../../routing/routers/Paths"
+import { usePlayers } from "../../../../hooks/api/usePlayers"
+import { useCurrentGame } from "../../../../hooks/api/useCurrentGame"
 import { observer } from "mobx-react-lite"
 import { Login } from "@mui/icons-material"
 import { Button } from "@mui/material"
-import BaseContainer from "../ui/BaseContainer"
-import "../../styles/ui/Divs.scss"
-import "../../styles/ui/Cards.scss"
-import { useCards } from "../../hooks/api/useCards"
-import ScoreAnnouncing from "../ui/ScoreAnnouncing"
-import { EntityModelCard } from "../../generated"
+import BaseContainer from "../../../ui/BaseContainer"
+import "../../../../styles/ui/Divs.scss"
+import "../../../../styles/ui/Cards.scss"
+import { useCards } from "../../../../hooks/api/useCards"
+import ScoreAnnouncing from "./ScoreAnnouncing"
+import { EntityModelCard } from "../../../../generated"
 import { toJS } from "mobx"
 
 /**
