+++ src/components/views/game/room/ControlPanel.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { useNavigate } from "react-router-dom"
-import { useCurrentGame } from "../../hooks/api/useCurrentGame"
-import { Paths } from "../routing/routers/Paths"
+import { useCurrentGame } from "../../../../hooks/api/useCurrentGame"
+import { Paths } from "../../../routing/routers/Paths"
 import Button from "@mui/material/Button"
 import SendIcon from "@mui/icons-material/Send"
 import LogoutIcon from "@mui/icons-material/Logout"
