+++ src/components/views/game/room/Room.scss
@@ -1,4 +1,4 @@
-@import '../../styles/theme';
+@import 'src/styles/theme';
 
 .room {
   padding: 0 20px;
