+++ src/components/views/game/room/Room.tsx
@@ -1,10 +1,10 @@
 import React from "react"
 import { observer } from "mobx-react-lite"
-import { useCurrentGame } from "../../hooks/api/useCurrentGame"
-import BaseContainer from "../ui/BaseContainer"
-import { RoomInfoContainer } from "../ui/RoomInfoContainer"
-import { ControlPanel } from "../ui/ControlPanel"
-import { RoomPlayerContainer } from "../ui/RoomPlayerContainer"
+import { useCurrentGame } from "../../../../hooks/api/useCurrentGame"
+import BaseContainer from "../../../ui/BaseContainer"
+import { RoomInfoContainer } from "./RoomInfoContainer"
+import { ControlPanel } from "./ControlPanel"
+import { RoomPlayerContainer } from "./RoomPlayerContainer"
 import { Grid } from "@mui/material"
 import "./Room.scss"
