+++ src/components/views/lobby/Lobby.tsx
@@ -1,19 +1,19 @@
 import React, { useEffect } from "react"
 import { observer } from "mobx-react-lite"
 import { toJS } from "mobx"
-import Footer from "../ui/Footer"
-import BaseContainer from "../ui/BaseContainer"
+import Footer from "../../ui/Footer"
+import BaseContainer from "../../ui/BaseContainer"
 import { useNavigate } from "react-router-dom"
-import { useGames } from "../../hooks/api/useGames"
-import { usePlayers } from "../../hooks/api/usePlayers"
-import { CreateGame } from "../ui/CreateGame"
-import { RoomRow } from "../ui/RoomRow"
-import { Paths } from "../routing/routers/Paths"
+import { useGames } from "../../../hooks/api/useGames"
+import { usePlayers } from "../../../hooks/api/usePlayers"
+import { CreateGame } from "./CreateGame"
+import { RoomRow } from "./RoomRow"
+import { Paths } from "../../routing/routers/Paths"
 import { Grid } from "@mui/material"
-import "../../styles/ui/Button.scss"
-import "../../styles/ui/Divs.scss"
-import "../../styles/ui/Lists.scss"
-import "../../styles/ui/images.scss"
+import "../../../styles/ui/Button.scss"
+import "../../../styles/ui/Divs.scss"
+import "../../../styles/ui/Lists.scss"
+import "../../../styles/ui/images.scss"
 
 const Lobby = observer(() => {
   const navigate = useNavigate()
